Add new function consolidate_safetensors_files_on_every_rank for HF consolidation
Currently we are only using rank-0 for HF consolidation. But we should be able to use every rank to consolidate the sharded files, which will speed up the consolidation by Nx (where N is the number of ranks). Adding a new method consolidate_safetensors_files_on_every_rank to do this.

Differential Revision: [D79000720](https://our.internmc.facebook.com/intern/diff/D79000720/)

ghstack-source-id: 299316617
Pull Request resolved: https://github.com/pytorch/pytorch/pull/159393